requirements.txt
pyyaml
numpy
similaritymeasures
matplotlib